feature(product) Fixing removing system attributes during import - fixes after CR
